undo resource-names move
