+++ libs/wingsdk/src/cloud/function.ts
@@ -3,7 +3,7 @@ import { join } from "path";
 import { Construct } from "constructs";
 import { fqnForType } from "../constants";
 import { App } from "../core/app";
-import { CaseConventions, ResourceNames } from "../core/resource-names";
+import { CaseConventions, ResourceNames } from "../shared/resource-names";
 import { Duration, IInflightHost, IResource, Resource } from "../std";
 
 /**
