+++ libs/wingsdk/src/shared/resource-names.ts
