+++ libs/wingsdk/src/target-tf-aws/function.ts
@@ -10,8 +10,8 @@ import { LambdaPermission } from "../.gen/providers/aws/lambda-permission";
 import { S3Object } from "../.gen/providers/aws/s3-object";
 import * as cloud from "../cloud";
 import * as core from "../core";
-import { NameOptions, ResourceNames } from "../core/resource-names";
 import { createBundle } from "../shared/bundling";
+import { NameOptions, ResourceNames } from "../shared/resource-names";
 import { PolicyStatement } from "../shared-aws";
 import { IInflightHost, Resource } from "../std";
 import { Duration } from "../std/duration";
