+++ libs/wingsdk/src/target-tf-aws/queue.ts
@@ -5,8 +5,8 @@ import { LambdaEventSourceMapping } from "../.gen/providers/aws/lambda-event-sou
 import { SqsQueue } from "../.gen/providers/aws/sqs-queue";
 import * as cloud from "../cloud";
 import * as core from "../core";
-import { NameOptions, ResourceNames } from "../core/resource-names";
 import { convertBetweenHandlers } from "../shared/convert";
+import { NameOptions, ResourceNames } from "../shared/resource-names";
 import { calculateQueuePermissions } from "../shared-aws/permissions";
 import { IInflightHost, Resource } from "../std";
