+++ libs/wingsdk/src/target-tf-aws/secret.ts
@@ -5,7 +5,7 @@ import { DataAwsSecretsmanagerSecret } from "../.gen/providers/aws/data-aws-secr
 import { SecretsmanagerSecret } from "../.gen/providers/aws/secretsmanager-secret";
 import * as cloud from "../cloud";
 import * as core from "../core";
-import { NameOptions, ResourceNames } from "../core/resource-names";
+import { NameOptions, ResourceNames } from "../shared/resource-names";
 import { calculateSecretPermissions } from "../shared-aws/permissions";
 import { IInflightHost } from "../std";
