+++ libs/wingsdk/src/target-tf-azure/bucket.ts
@@ -17,7 +17,7 @@ import {
   CaseConventions,
   NameOptions,
   ResourceNames,
-} from "../core/resource-names";
+} from "../shared/resource-names";
 import { IInflightHost } from "../std";
 
 /**
